Rank class 생성
rank의 정보를 담는 Rank class를 생성했다.
